EEdge has its own integration tests now
