+++ src/python/engagement_edge/tests/test_engagement_edge.py
@@ -1,8 +1,9 @@
+import pytest
 import unittest
 
 from chalice.test import Client
 from src.engagement_edge import JWT_SECRET, app
-from src.lib.sagemaker import create_sagemaker_client
+from grapl_tests_common.clients.engagement_edge_client import EngagementEdgeClient
 
 # gross hack because engagement edge is pseudo singleton
 JWT_SECRET.secret = "hey im a fake secret"
@@ -10,19 +11,25 @@ JWT_SECRET.secret = "hey im a fake secret"
 
 class TestEngagementEdgeChalice(unittest.TestCase):
     def test_requires_auth_fails_without_cookie_headers(self):
-        with Client(app) as cient:
+        with Client(app) as client:
             result = client.http.post(
                 "/getNotebook",
                 headers={
                     "Origin": "https://local-grapl-engagement-ux-bucket.s3.amazonaws.com"
                 },
             )
-            result = client.http.post("/getNotebook")
             assert result.status_code == 400
             assert result.json_body == {"error": "Must log in"}
 
 
-class TestSagemakerClient(unittest.TestCase):
-    def test_when_is_local_returns_8888(self):
-        client = create_sagemaker_client(is_local=True)
-        assert client.get_presigned_url() == "http://localost:8888/"
+@pytest.mark.integration_test
+class TestEngagementEdgeClient(unittest.TestCase):
+    def test_get_jwt(self) -> None:
+        client = EngagementEdgeClient(use_docker_links=True)
+        client.get_jwt()
+
+    def test_get_notebook_link(self) -> None:
+        client = EngagementEdgeClient(use_docker_links=True)
+        jwt = client.get_jwt()
+        notebook_url = client.get_notebook(jwt=jwt)
+        assert "localhost:8888" in notebook_url
